PG-1970: Improve output on failure
If we detect a scope mismatch we print both the received scopes and the
required scopes (from pg_hba) for easier debugging.

Additionally we print out the username with DEBUG1, in case somebody has
issues logging in even with correct scopes. This allows easier debugging
to verify that the map file is correct and/or the user exists in
postgres.