PG-1967: Basic keycloak test setup on CI
This commit adds a very basic test using keycloak + rpec.

* Keycloak is ran using docker, with a previously generated
  configuration which just gets imported
* The test is written in rspec, for easy extensibility (different
  providers, different testcases)
* For now it's just a simple smoke test to verify that we can log in
  with the validator

One particularly ugly part of this is that we install the seflf signed
certificates systemwide, but the only other option is using
`OAUTH_DEBUG=unsafe`, which changes other behavior and results in tons
of debug output. Trusting this certificate by the system seems like a
more-realstic approach.